knodb, runtime, engine, coverage updates
* KNAPP: Fixes to module logic
* KNODB: Add knodb/altpath knodb/bakpath, improve packpool messaging
* KNODB: Add functions for getting hashsets of keys from various kinds of hashindexes
* KNODB: Move flexsplit to flexpack
* MISC: Add some indentation rules to knoc.el
* COVERAGE: Don't use profiling settings for coverage
* COVERAGE:  update coverage makefile targets
* COVERAGE: Update gcov config to include latest lcov
* COVERAGE: Various fixes and code cleanup
* COVERAGE: Fix choiceizeprim profiling case
* COVERAGE: Remove libtest runs from coverage tests
* STORAGE: Improve prefetch error checking
* STORAGE: Remove unused code and covers_slotids field on indexes
* OPTIMIZE: improve logging
* RUNTIME: Reimplement hashset/vector functions, add hashset xtypes
* RUNTIME:ERRORS: Fill in details from irritant in some cases
* SCHEME: Fix double eval in TRYIF
* SCHEME: Add more conditional cases which return {} on {} tests
* ENGINE: General Improvements to engine implementation, handle
  1-thread case by not using threads at all
* MODULES: Fix bad signature in ezrecord macro implementation
* TESTS: Add optmode and tryif tests
* PACKAGING: Update .zip build targets

+++ include/kno/tables.h
@@ -713,7 +713,7 @@ KNO_EXPORT int kno_hashset_add(kno_hashset h,lispval keys);
 KNO_EXPORT lispval kno_hashset_intern(kno_hashset h,lispval key,int add);
 
 KNO_EXPORT lispval kno_hashset_elts(kno_hashset h,int clean);
-KNO_EXPORT lispval *kno_hashset_vec(struct KNO_HASHSET *h,int *len);
+KNO_EXPORT lispval *kno_hashset_vec(struct KNO_HASHSET *h,ssize_t *len);
 
 KNO_EXPORT void kno_init_hashset(kno_hashset h,int n,int stack_cons);
 KNO_EXPORT lispval kno_make_hashset(void);
